// client/dist/widgets/common/search/src/config.ts
export enum SearchResultView {
  ResultPanel = 'Result_Panel',
  OtherWidgets = 'Other_Widgets'
}

export enum ArrangementStyle {
  Style1 = 'Style1',
  Style2 = 'Style2'
}

export enum SearchResultStyle {
  Classic = 'Classic',
  Compact = 'Compact'
}
// ...
export interface DatasourceList {
  [key: string]: DatasourceListItem
}

export interface GeocodeList {
  [key: string]: GeocodeListItem
}

export type ServiceList = DatasourceList & GeocodeList

export type ServiceListItem = DatasourceListItem & GeocodeListItem

export type IMServiceListItem = ImmutableObject<ServiceListItem>

export interface DatasourceSQLList {
  [key: string]: SqlExpression[]
}
export type IMDatasourceSQLList = ImmutableObject<DatasourceSQLList>

export interface DatasourceListItem {
  searchServiceType: SearchServiceType
  maxSuggestions: number
  resultMaxNumber: number
  useDataSource: UseDataSource
  displayFields?: FieldSchema[]
  searchFields?: FieldSchema[]
  SQL?: SqlResult
  SuggestionSQL?: SqlResult
  searchExact?: boolean
  hint?: string
  isFromSuggestion?: boolean
  configId: string
  searchText?: string
}

export interface GeocodeListItem {
  searchServiceType: SearchServiceType
  maxSuggestions: number
  resultMaxNumber: number
  label: string
  geocodeURL: string
  outputDataSourceId: string
  hint?: string
  icon?: any
  magicKey?: string
  configId: string
  singleLineFieldName?: string
  displayFields?: FieldSchema[]
  defaultAddressFieldName?: string
  addressFields?: FieldSchema[]
  isSupportSuggest?: boolean
  searchText?: string
}

export type IMDatasourceList = ImmutableObject<DatasourceList>
export type IMGeocodeList = ImmutableObject<GeocodeList>
export type IMServiceList = ImmutableObject<ServiceList>
export type IMDatasourceListItem = ImmutableObject<DatasourceListItem>

export interface config extends SearchSuggestionConfig {
  datasourceConfig: SearchDataConfig[]
  hint: string
  searchResultView: SearchResultView
  arrangementStyle: ArrangementStyle
  searchResultStyle: SearchResultStyle
  resultMaxNumber: number
  isAutoSelectFirstResult: boolean
  // link
  linkParam?: IMLinkParam
}

export type IMConfig = ImmutableObject<config>